Added The Shop Menu
The shop menu has full functionality. I also added the text on the left side that states how much of each currency you have, NO FUNCTIONALITY. Next I made a new script that we are going to use to put all the prefab references. I did this because I doing it all in the placing file is bad practice, and would make the code harder to read in my opinion. I also implemented text mesh pro if who ever is reading this, is wondering why there are emojis on the changed file Xd.